test(architecture): move final admin root tests
