+++ docs/implementation/test-arch-55-admin-root-final.md
@@ -1,0 +1,7 @@
+# TEST-ARCH-55 - Admin root final cleanup
+
+Moved the final two admin root tests into structured test directories.
+
+Updated only test-relative imports after relocation.
+
+Guardrails: no runtime/product/API/auth/DB/schema/migration/dependency/lockfile/CI changes.
